% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccineSurveillanceMethodEvaluation.R
\docType{package}
\name{Eumaeus-package}
\alias{Eumaeus}
\alias{Eumaeus-package}
\title{Eumaeus: Evaluating Use of Methods for Adverse Event Under Surveillance}
\description{
Similar to our previous research on the evaluation of causal effect estimation methods, this study aims to systematically evaluate methods for (safety) surveillance of vaccines.
}
\keyword{internal}
