% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  databaseId,
  databaseName = databaseId,
  databaseDescription = databaseId,
  minCellCount = 5,
  maxCores = 1,
  exposureIds = getExposuresOfInterest()$exposureId,
  verifyDependencies = TRUE,
  createCohorts = TRUE,
  synthesizePositiveControls = TRUE,
  runCohortMethod = TRUE,
  runSccs = TRUE,
  runCaseControl = TRUE,
  runHistoricalComparator = TRUE,
  computeCriticalValues = TRUE,
  createDbCharacterization = TRUE,
  exportResults = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where outcome data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the \code{cohortDatabaseSchema}.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{exposureIds}{The IDs of the exposure cohorts to include. See \code{\link{getExposuresOfInterest}}
for the available exposure cohorts and their IDs.}

\item{verifyDependencies}{Check whether correct package versions are installed?}

\item{createCohorts}{Create the exposure and outcome cohorts?}

\item{synthesizePositiveControls}{Should positive controls be synthesized?}

\item{runCohortMethod}{Perform the cohort method analyses?}

\item{runSccs}{Perform the SCCS and SCRI analyses?}

\item{runCaseControl}{Perform the case-control analyses?}

\item{runHistoricalComparator}{Perform the historical comparator analyses?}

\item{computeCriticalValues}{Compute critical values for all methods?}

\item{createDbCharacterization}{Create a high-level characterization of the database?}

\item{exportResults}{Export the results to a single zip file (containing several CSV files) for sharing?}
}
\description{
Execute the Study
}
\details{
This function executes the Study.
}
