% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{getExposuresOfInterest}
\alias{getExposuresOfInterest}
\title{Get the exposures of interest}
\usage{
getExposuresOfInterest()
}
\value{
A tibble with exposure IDs and names.
}
\description{
Get the exposures of interest
}
