% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UploadResultsToDatabase.R
\name{getResultsDataModelSpecifications}
\alias{getResultsDataModelSpecifications}
\title{Get specifications for the Eumaeus results data model}
\usage{
getResultsDataModelSpecifications()
}
\value{
A tibble data frame object with specifications
}
\description{
Get specifications for the Eumaeus results data model
}
