% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortDiagnosics.R
\name{runCohortDiagnostics}
\alias{runCohortDiagnostics}
\title{Execute the Feasibility assessment}
\usage{
runCohortDiagnostics(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable,
  outputFolder,
  minCellCount,
  databaseId,
  databaseName = databaseId,
  databaseDescription = databaseId,
  createCohorts = TRUE,
  runCohortDiagnostics = TRUE
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where outcome data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the \code{cohortDatabaseSchema}.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{minCellCount}{The minimum cell count for fields contains person counts or fractions.}

\item{databaseId}{A short string for identifying the database (e.g. 'Synpuf').}

\item{databaseName}{The full name of the database.}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{createCohorts}{Create the exposure and outcome?}

\item{runCohortDiagnostics}{Run cohort diagnostics?}
}
\description{
Execute the Feasibility assessment
}
\details{
This function executes the Study.

The \code{createCohorts}, \code{synthesizePositiveControls}, \code{runAnalyses}, and \code{runDiagnostics} arguments
are intended to be used to run parts of the full study at a time, but none of the parts are considerd to be optional.
}
